# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MorphieCore minimal model interaction gateway",
    "version": "0.1.0",
    "description": "Current envelope-level HTTP contract, not complete standard conformance. Responses uses the Standard adapter; Chat retains the explicit MorphieCore compatibility profile. Provider-specific rules cannot authorize private Responses carriers or silent loss. The independent _openbridge attachment is rejected on protocol envelopes/items and never emitted. Static Images, binary Speech and file Transcription are independent tasks with explicit single-source API-key activation, never inferred from conversation bindings. Binary media entries require explicit model selection and the matching single-source, non-fallback credential pool; embedded callers also activate entries explicitly. Named ancillary audio-report projections preserve full IR reports while emitting standard audio/JSON. Models exposes only activated public labels; fine-tuned model deletion is not authorized. No stateful resources, URL fetching or implicit local retry/fallback. Realtime and downstream audio SSE are not admitted."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "security": [{"bearerAuth": []}],
  "paths": {
    "/v1/models": {
      "get": {
        "operationId": "listModels",
        "summary": "List activated task public labels",
        "description": "Immutable local view, deduplicated across client protocols and candidates and sorted by label. No pagination, upstream discovery, capability/price extensions or request body. HEAD shares this route without a response body.",
        "responses": {
          "200": {
            "description": "Complete standard model list, bounded at startup; never a truncated successful prefix.",
            "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ModelList"}}}
          },
          "401": {"$ref": "#/components/responses/Error"},
          "405": {"$ref": "#/components/responses/Error"},
          "503": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/v1/models/{model}": {
      "parameters": [{
        "in": "path", "name": "model", "required": true,
        "schema": {"$ref": "#/components/schemas/PublicModelLabel"}
      }],
      "get": {
        "operationId": "retrieveModel",
        "summary": "Retrieve an activated public model",
        "responses": {
          "200": {
            "description": "The same model object as in the list; no Provider or account information.",
            "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Model"}}}
          },
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"},
          "405": {"$ref": "#/components/responses/Error"},
          "503": {"$ref": "#/components/responses/Error"}
        }
      },
      "delete": {
        "operationId": "deleteModel",
        "summary": "Reject deletion without fine-tuned model ownership",
        "description": "This instance has no owned fine-tuned models or deletion capability. Activated labels return model_deletion_forbidden (403); unknown or inactive labels return model_not_found (404). No registry/activation mutation, upstream request or fabricated deleted:true response.",
        "responses": {
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"},
          "403": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"},
          "503": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "operationId": "chatCompletion",
        "parameters": [{"$ref": "#/components/parameters/ProbeId"}],
        "summary": "Generate text from admitted text or user image input through single-candidate Chat",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatRequest"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/Generation"},
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"},
          "405": {"$ref": "#/components/responses/Error"},
          "408": {"$ref": "#/components/responses/Error"},
          "413": {"$ref": "#/components/responses/Error"},
          "415": {"$ref": "#/components/responses/Error"},
          "429": {"$ref": "#/components/responses/Error"},
          "500": {"$ref": "#/components/responses/Error"},
          "502": {"$ref": "#/components/responses/Error"},
          "503": {"$ref": "#/components/responses/Error"},
          "504": {"$ref": "#/components/responses/Error"}
        }
      }
    },
...
    "/v1/responses": {
      "post": {
        "operationId": "response",
        "parameters": [{"$ref": "#/components/parameters/ProbeId"}],
        "summary": "Generate text from explicitly admitted text, image, inline or URL file input through stateless Responses",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponsesRequest"}}}
        },
        "responses": {
          "200": {"$ref": "#/components/responses/Generation"},
          "400": {"$ref": "#/components/responses/Error"},
          "401": {"$ref": "#/components/responses/Error"},
          "404": {"$ref": "#/components/responses/Error"},
          "405": {"$ref": "#/components/responses/Error"},
          "408": {"$ref": "#/components/responses/Error"},
          "413": {"$ref": "#/components/responses/Error"},
          "415": {"$ref": "#/components/responses/Error"},
          "429": {"$ref": "#/components/responses/Error"},
          "500": {"$ref": "#/components/responses/Error"},
          "502": {"$ref": "#/components/responses/Error"},
          "503": {"$ref": "#/components/responses/Error"},
          "504": {"$ref": "#/components/responses/Error"}
        }
      }
    }
  },
  "components": {
    "parameters": {
      "ProbeId": {
        "in": "header", "name": "x-morphiecore-probe-id", "required": false,
        "schema": {"type": "string"},
        "description": "Operator correlation only when private probe diagnostics are enabled and authentication succeeds. Valid form: 32 lowercase hex run ID, colon, then decimal 1..999999. Invalid/duplicate values disable observation only, not business admission. Never forwarded upstream or echoed downstream. Not semantic request data."
      }
    },
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "description": "Exactly one Authorization header. Configured startup secret, not an upstream API key."}},
    "schemas": {
      "Model": {
        "type": "object", "required": ["id", "object", "created", "owned_by"],
        "additionalProperties": false,
        "properties": {
          "id": {"$ref": "#/components/schemas/PublicModelLabel"},
          "object": {"const": "model"},
          "created": {
            "type": "integer", "format": "unixtime", "minimum": 1, "maximum": 9223372036854775807,
            "description": "Developer publication time in Unix seconds. Date-only sources use UTC midnight; not a router listing or Gateway startup timestamp."
          },
          "owned_by": {"type": "string", "minLength": 1, "maxLength": 128, "description": "Model developer name, not inference Provider or account owner."},
          "shutdown_date": {"type": ["string", "null"], "format": "date", "description": "Optional standard report; omitted when unreported, not fabricated."}
        }
      },
      "ModelList": {
        "type": "object", "required": ["object", "data"], "additionalProperties": false,
        "properties": {
          "object": {"const": "list"},
          "data": {"type": "array", "maxItems": 64, "items": {"$ref": "#/components/schemas/Model"}}
        }
      },
...
      "PublicModelLabel": {
        "type": "string", "minLength": 1, "maxLength": 64, "pattern": "^[a-z0-9_.-]+$",
        "description": "Activated public model label without a provider/ prefix. Resolves a fixed trusted Route; canonical identity and Provider-specific upstream model names remain internal. Clients cannot select a Provider."
      },
      "CacheSession": {
        "type": "string", "minLength": 1, "maxLength": 256, "pattern": "^[^\\u0000-\\u001f\\u007f-\\u009f]+$",
        "description": "MorphieCore scoped extension: caller-supplied Provider cache/observability grouping, not a gateway-managed session. Only explicitly supporting targets may receive it. No derivation from prompt_cache_key, user, safety_identifier or opaque replay; unsupported targets reject it. OpenCode Go maps an explicit ASCII session_id to x-opencode-session instead of the upstream body; absent values do not create an identity. Incoming session headers are not forwarded."
      },
      "ChatRequest": {
        "type": "object", "required": ["model", "messages"],
        "not": {"anyOf": [{"required": ["provider"]}, {"required": ["_openbridge"]}]},
        "properties": {
          "model": {"$ref": "#/components/schemas/PublicModelLabel"},
          "session_id": {"$ref": "#/components/schemas/CacheSession"},
          "prompt_cache_key": {"type": ["string", "null"], "maxLength": 256, "description": "Advisory Provider cache-affinity key. Targets without a declared key carrier omit it; automatic caching is independent. Does not select or reorder MorphieCore candidates."},
          "messages": {"type": "array", "items": {"type": "object", "not": {"required": ["_openbridge"]}}, "description": "User/assistant and system/developer content accepts a string or nonempty ordered plain-text part array. Tool result content accepts a string or ordered plain-text part array, including an empty array; single tool-result parts remain arrays. User content additionally admits ordered text/image_url parts when public and endpoint contracts permit image input; one image remains an array. HTTP(S) URL and Base64 image data URL only; file IDs, other roles' images, tool-result media remains rejected. Assistant audio history has separate library admission and trusted source/expiry requirements in the Chat media profile; unbound audio IDs and existing text/image-only product bindings reject it."},
          "modalities": {"type": ["array", "null"], "items": {"enum": ["text", "audio"]}, "minItems": 1, "maxItems": 2, "uniqueItems": true, "description": "Library output intent; audio requires matching audio options and explicit public/endpoint audio_output admission. Existing text/image bindings do not gain audio admission."},
          "audio": {"type": ["object", "null"], "required": ["format", "voice"], "additionalProperties": false, "properties": {"format": {"enum": ["wav", "aac", "mp3", "flac", "opus", "pcm16"]}, "voice": {"type": "string", "minLength": 1, "maxLength": 128}}, "description": "Named-voice output options, not voice/account management. Requires audio modality; streaming requires pcm16. No audio model is activated by this schema."},
          "stream": {"type": ["boolean", "null"], "default": false},
          "max_completion_tokens": {"type": ["integer", "null"], "minimum": 1, "description": "Bootstrap defaults absent/null to 1024 only when the public contract admits this control, and rejects values above 16384 after semantic decode. Targets without output-token control do not receive a fabricated cap."},
          "metadata": {"type": ["object", "null"], "maxProperties": 16, "propertyNames": {"maxLength": 64}, "additionalProperties": {"type": "string", "maxLength": 512}, "description": "Typed request hint, never copied into reported response metadata."},
          "service_tier": {"type": ["string", "null"], "enum": ["auto", "default", "flex", "scale", "priority", "fast", null], "description": "Request hint; the reported effective tier may differ. Endpoint context admission remains independent."},
          "logprobs": {"type": ["boolean", "null"], "description": "Return token probabilities when true; null/false are inactive controls."},
          "top_logprobs": {"type": ["integer", "null"], "minimum": 0, "maximum": 20, "description": "A non-null integer requires logprobs:true; null normalizes to absent."}
        },
        "allOf": [{"if": {"required": ["top_logprobs"], "properties": {"top_logprobs": {"type": "integer"}}}, "then": {"required": ["logprobs"], "properties": {"logprobs": {"const": true}}}}],
        "additionalProperties": true,
        "description": "Additional fields require admission by the Chat text profile and selected public contract; additionalProperties does not authorize arbitrary upstream parameters. Function-only tool_choice.allowed_tools uses the nested Chat shell. Explicit assistant-message/tool-call membership cannot be projected to a standard Responses target; incompatible history is rejected before upstream I/O, without dropping the owner or membership. MorphieCore's Chat extension admits scoped openai-responses-v1 reasoning_details on assistant history; preserve those actual output details for replay rather than treating them as ordinary assistant text."
      },
...
      "Error": {
        "type": "object", "required": ["error"], "additionalProperties": false,
        "properties": {"error": {
          "type": "object", "required": ["message", "type", "param", "code"], "additionalProperties": false,
          "properties": {"message": {"type": "string"}, "type": {"type": "string", "enum": ["invalid_request_error", "api_error"]}, "param": {"type": "null"}, "code": {"type": "string"}}
        }}
      }
    },
    "responses": {
      "Generation": {
        "description": "Current profile-specific JSON response or incremental SSE, retaining the public model label. Typed output, usage and dependencies must be representable by the selected client contract; unrepresentable facts are not silently dropped. No independent _openbridge attachment is emitted. Other named profile differences still prevent a full standard-conformance claim. Exact text, tool, usage, grouping and replay admission belongs to docs/architecture/responses-text-profile.md, chat-text-profile.md and client-generation-profile.md. Chat audio/citations have separate library scope in chat-media-profile.md, not automatic product activation. HTTP 200 does not prove semantic completion, stream closure or output adherence. Late failures abort the body without a fabricated successful terminal.",
        "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}},
        "content": {"application/json": {"schema": {"type": "object", "not": {"required": ["_openbridge"]}}}, "text/event-stream": {"schema": {"type": "string"}}}
      },
      "Error": {
        "description": "Sanitized application error. 401 includes WWW-Authenticate: Bearer. No upstream body, origin, credential locator or parser input is reflected. HTTP parser failures occur outside this envelope.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      }
    }
  }
}